18/09/2025 p1
new bool to disable camera -isRecognising-,
function enableCamera sets isRecognising to true and sets signatures to 0

stored the Y coordinate of the parking in case it be randomized with parkingY variable and saveParkingPosition function
pitiditos delays reduced
we set the default path of the first lane to be the outer, to avoid detecting the parking lot by looking away from it
reduced the frequency of calling the automovecamera function
when disabling the camera after the first lap, the camera is set to 0 only once, not repeteadly
changed reposicionar to be able to check if the distance0 is measured properly
turned camera operation off after the first tramo of the 2nd lap
the camera now stops tracking the blocks before changing tramo, disabling camera during the process to avoid errors. camera is then enabled back on every turn and every tramo change.
new define: width of the track lanes