Changes from the National Final
Increased speed during the recognition lap to avoid getting stuck.
Increased KD and KP for better turns.
Corrected small errors of the autoMoveCamera in the first and last tramos.
added the praking stop Y for anticlockwise laps
updated the LED and buzzer pins for the new PCB in the ESP32 Master
deleted the Melody player since now we use a buzzer not a speaker
increased the KP and KD in the Master for sharper turns
reset reason sent at setup
deleted debug LED and buzzer calls for the competition