Shared libraries
The projects now import the common libraries from the libraries folder in code.
The corresponding libraries have been removed from the projects.
Platformio.ini were modified to account for this change.
In the pin assignments libraries, an include folder was created so that platform IO compiler could find the header files.