LIDAR Testing v1
FIrst time Unit Testing in Platform IO.
Created test file to test the ransac algorithm code in the native enviroment. Platform.ini updated to add native enviroment.
Created lidar Handling to store the main functions to deal with the LIDAR. lidarMeasurement struct created to store range and bearing. Created a lidar storage class to easily store the data received from the LIDAR, and convert it to cartesian coordinates. Since it handles an array, it has an explicitly defined class destructor. Added the LIDAR Task code for future development.
Created a statistics.h to calculate statistical variables which will probably not be used in the future.
Ransac algorithm file created. Created types Poin2D and Line2D. Implemented functions to calculate the perpendicular point on a line, 2D distance, 2D squared distance, the line of best fit given a point set and to execute the RANSAC algorithm.
Small changes made to LIDAR Tests/main. created a lidarBuffer to receive the data and record the number of measurements received in measurementIndex. Startbir recorded.