Lidar testing v2
Added that the ransac algorithm can now detect multiple lines, and return the number of lines detected.
Added that if the number of unassociated points is less than the concensus, the algorithm should stop.
Created a calculate line of best fit from a sample of the data points given by an array with the indices of the points in the sample.
Corrected error with the trigonometry in the convert to cartesian function
Created a List class modifying a vector for its use in the ransac algorithm, where we want to exclude the inliers.
Changed unsigned ints to uint16_t for memory efficiency.
Added tests for the list class and added more points to the ransac test