RANSAC extraction completed (250ms)
The exeRANSAC function succesfully detects wall lines even when blocks are placed. It takes around 250ms to run. We might want to change the doubles to floats for space optimisation. This is intended to be improved.
Created a lidarDataTypes file to store the dataTypes used for the lidar management
Added a getMeasurement function to the lidarStorage to get a measurement directly from the dataBuffer. Added also a getSize method. also added a function to set the dataArray to a given array (using memcpy)
added a MIN SAMPLE define to reject samples too small. fused the distance2D and perpPointOnLine fro easier syntax.
calcBestFitLineFromSample now gets a reference of a list with the data indices.
Generated a seed for the random number generator with the random device when using the native enviroment and with esp_random when running in the esp32.
The RANSAC sample is now generated randomly with angles within DEGREE_RANGE from the centre point.
Added printf lines to debug.
Optimised and corrected the pop of points lying
Added an esp32 testing enviroment extending the normal esp32 enviroment in the pio.ini
Added test function to the native enviroment and created a test ransac for the esp32
Created a blockPositions array to store the positions of the blocks. To be worked on.
readDistance executes the RANSAC algorithm on the current data and returns the distance to the 0th line. Should give you the distance to the line whose bearing is in a range of the one requested.
The LIDAR task stores the measurements in a buffer. When a new scan begins we store the previous buffer in the lidar storage.
Created a file to store real lidar data in order to use the test_lidar_ransac_from_file