Python telemetry viewer
Created a telemetry manager in python that receives and displays the telemetry data. To execute it, run mainRoot.py
There are displays for the raw data, the lidar data, the camera data, and the car's trajectory.
There is yet work to be done. The displays are laggy, the interfaces are not polished and save button doesn't work yet.